+++ core/dbt/include/global_project/macros/materializations/incremental/on_schema_change.sql
@@ -57,7 +57,7 @@
   {%- set target_columns = adapter.get_columns_in_relation(target_relation) -%}
   {%- set source_not_in_target = diff_columns(source_columns, target_columns) -%}
   {%- set target_not_in_source = diff_columns(target_columns, source_columns) -%}
-  
+
   {% set new_target_types = diff_column_data_types(source_columns, target_columns) %}
 
   {% if source_not_in_target != [] %}
@@ -72,6 +72,8 @@
     'schema_changed': schema_changed,
     'source_not_in_target': source_not_in_target,
     'target_not_in_source': target_not_in_source,
+    'source_columns': source_columns,
+    'target_columns': target_columns,
     'new_target_types': new_target_types
   } %}
 
@@ -132,7 +134,11 @@
 
 {% macro process_schema_changes(on_schema_change, source_relation, target_relation) %}
     
-    {% if on_schema_change != 'ignore' %}
+    {% if on_schema_change == 'ignore' %}
+
+     {{ return({}) }}
+
+    {% else %}
     
       {% set schema_changes_dict = check_for_schema_changes(source_relation, target_relation) %}
       
@@ -158,6 +164,8 @@
         {% endif %}
       
       {% endif %}
+
+      {{ return(schema_changes_dict['source_columns']) }}
     
     {% endif %}